refactor(racingcar): 사용자 입력값 검증 로직 컨트롤러로 이전
