+++ src/main/kotlin/racingcar/constant/ErrorMessage.kt
@@ -4,3 +4,4 @@ const val ERROR_NAME_LENGTH = "이름의 길이가 올바르지 않습니다."
 const val ERROR_WRONG_NUMBER = "숫자가 아니거나 음수가 입력되었습니다."
 const val ERROR_INPUT_HANDLER = "입력값 검증과정에 문제가 있습니다."
 const val ERROR_CARS_CREATE = "차량생성중 들어온 이름 리스트가 빈값 혹은 NULL입니다"
+const val ERROR_WRONG_INPUT = "올바르지 않은 입력값입니다."
